+++ Net_Core/Web.Angular/projects/gw-lib/core/account/src/c/account-details/account-details.component.ts
@@ -1,10 +1,9 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
 
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { Subscription } from 'rxjs';
 // Angular Material
 import { MatButtonModule } from '@angular/material/button';
 import { MatCheckboxModule } from '@angular/material/checkbox';
@@ -50,12 +49,11 @@ import { AccountService } from '../../account.service';
   templateUrl: './account-details.component.html',
   styleUrls: ['./account-details.component.scss']
 })
-export class AccountDetailsComponent implements OnDestroy, OnInit {
+export class AccountDetailsComponent implements OnInit {
   private _AccountProfile!: IAccountProfile;
   private _SecurityInfoAccount: null | ISecurityInfo = null;
   private _SecurityInfoGroups: null | ISecurityInfo = null;
   private _SecurityInfoRoles: null | ISecurityInfo = null;
-  private _Subscription: Subscription = new Subscription();
 
   frmAccount!: FormGroup;
 
@@ -78,6 +76,8 @@ export class AccountDetailsComponent implements OnDestroy, OnInit {
   litLastNameWarning: string = '';
   litStatusWarning: string = '';
 
+  rolesAvailable: Array<string> = ['one', 'two'];
+  rolesSelected: Array<string> = [];
   rolesPickListName: string = 'roles';
 
   selectedStatus: number = 0;
@@ -132,11 +132,6 @@ export class AccountDetailsComponent implements OnDestroy, OnInit {
 		private _SecuritySvc: SecurityService
   ) { }
 
-  ngOnDestroy(): void {
-    // console.log('AccountDetailsComponent.ngOnDestroy called', this._AccountProfile);
-    this._Subscription.unsubscribe();
-  }
-
   ngOnInit(): void {
     let mDesiredAccount: string = '';
     /*eslint indent: ["error", 2, { "SwitchCase": 1 }]*/
@@ -164,19 +159,15 @@ export class AccountDetailsComponent implements OnDestroy, OnInit {
         break;
     }
     this._GroupSvc.getGroups().then((groups) => {                         // Request and Response Handler #1 (getGroups)
-      // TODO: this would indicate that the pick-list component isn't loaded at this point
-      // and we are simply adding a delay to give it time... need to find a better way
-      // such as a different lifecycle hook?
       if(groups != null) {
-        setTimeout(() => { this._DataSvc.notifyDataChanged(this.groupsPickListName + '_AvailableItems', groups); }, 500);
+        this.groupsAvailable = groups;
       }
       return this._RoleSvc.getRoles();                                    // Request #2 (getRoles)
     }).catch((error) => {                                                 // Error Handler #1 (getGroups)
       this._LoggingSvc.toast('Error getting groups:\r\n' + error, 'Account Details:', LogLevel.Error);
     }).then((roles) => {                                                  // Response Handler #2 (getRoles)
       if(roles != null) {
-        // TODO: Same issue as Request #1
-        setTimeout(() => { this._DataSvc.notifyDataChanged(this.rolesPickListName + '_AvailableItems', roles); }, 500);
+        this.rolesAvailable = roles;
       }
       return this._SecuritySvc.getSecurityInfo('Accounts');               // Request #3 (getSecurityInfo('Accounts'))
     }).catch((error) => {                                                 // Error Handler #2 (getRoles)
@@ -208,13 +199,12 @@ export class AccountDetailsComponent implements OnDestroy, OnInit {
         let mRoles: string[] = [];
         let mGroups: string[] = [];
         this.isSysAdmin = this._AccountProfile.isSystemAdmin;
-        if(!this._GWCommon.isNullOrUndefined(this._AccountProfile.assignedRoles)) {
-          mRoles = this._AccountProfile.assignedRoles!;
-          setTimeout(() => { this._DataSvc.notifyDataChanged(this.rolesPickListName + '_SelectedItems', mRoles); }, 500);
+        if(this._AccountProfile && this._AccountProfile.assignedRoles) {
+          this.rolesSelected = this._AccountProfile.assignedRoles;
         }
-        if(!this._GWCommon.isNullOrUndefined(this._AccountProfile.groups)) {
+        if(this._AccountProfile && this._AccountProfile.groups) {
           mGroups = this._AccountProfile.groups!;
-          setTimeout(() => { this._DataSvc.notifyDataChanged(this.groupsPickListName + '_SelectedItems', mGroups); }, 500);
+          this.groupsSelected = mGroups;
         }
         this.applySecurity();
         this.populateForm();
@@ -350,7 +340,7 @@ export class AccountDetailsComponent implements OnDestroy, OnInit {
           this._LoggingSvc.toast('Error saving account!', 'Save Account', LogLevel.Error);
         });          
       } else {
-        console.log('AccountProfile', this._AccountProfile);
+        // console.log('AccountProfile', this._AccountProfile);
         this._AccountSvc.registerAccount(this._AccountProfile).then((mMessage: string) => {
           this._LoggingSvc.toast(mMessage, 'Register Account', LogLevel.Success);
           this.closeModal();
@@ -422,12 +412,12 @@ export class AccountDetailsComponent implements OnDestroy, OnInit {
     if(this.isRegistration) {
       this._AccountProfile.account = this.controls['email'].getRawValue();
     }
-    // this._AccountProfile.assignedRoles = '';
+    this._AccountProfile.assignedRoles = JSON.parse(JSON.stringify(this.rolesSelected));
     this._AccountProfile.email = this.controls['email'].getRawValue();
     this._AccountProfile.enableNotifications = this.controls['enableNotifications'].getRawValue();
     this._AccountProfile.failedAttempts = this.controls['failedAttempts'].getRawValue();
     this._AccountProfile.firstName = this.controls['firstName'].getRawValue();
-    // this._AccountProfile.groups = '';
+    this._AccountProfile.groups = JSON.parse(JSON.stringify(this.groupsSelected));
     this._AccountProfile.isSystemAdmin = this.controls['isSystemAdmin'].getRawValue();
     this._AccountProfile.lastName = this.controls['lastName'].getRawValue();
     this._AccountProfile.location = this.controls['location'].getRawValue();